requirements.txt
alembic==1.17.2
Authlib==1.6.5
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.12.0
crypto==1.4.1
cryptography==46.0.3
dnspython==2.8.0
dominate==2.9.1
email-validator==2.3.0
flake8==7.3.0
Flask==3.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
infinity==1.5
iniconfig==2.3.0
intervals==0.9.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
ms-identity-python @ git+https://github.com/azure-samples/ms-identity-python@834583646df315b79bc89f54808f4ae9c2a1ce59
msal==1.34.0
msgspec==0.20.0
Naked==0.1.32
packaging==25.0
pbkdf2==1.3
phonenumberslite==9.0.19
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
shellescape==3.8.1
six==1.17.0
SQLAlchemy==2.0.44
SQLAlchemy-Utils==0.42.0
srp==1.0.22
types-urllib3==1.26.25.14
typing_extensions==4.15.0
urllib3==2.5.0
validators==0.35.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-Alchemy==0.19.1
WTForms-Components==0.11.0
WTForms-SQLAlchemy==0.4.2

pywebpush==2.1.2
aiohttp==3.13.2
http-ece==1.2.1
py-vapid==1.9.2
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
attrs==25.4.0
frozenlist==1.8.0
multidict==6.7.0
propcache==0.4.1
yarl==1.22.0
psycopg2-binary
